backend/app/requirements.txt
fastapi==0.115.0
fastapi-jwt[authlib]==0.3.0
uvicorn==0.31.0
gunicorn==21.0.0
python-multipart==0.0.11
requests==2.32.3
dotenv==0.9.9
httpx==0.28.1
xgboost==3.0.2
beartype==0.21.0
python-docx==1.2.0
pdfplumber==0.11.7
pypdf==5.6.0
huggingface_hub==0.33.0
six==1.17.0
opencv-python==4.11.0.86
onnxruntime==1.22.0
shapely==2.1.1
pyclipper==1.3.0.post6
datrie==0.8.2
hanziconv==0.3.2
nltk==3.9.1
chardet==5.2.0
tiktoken==0.9.0
pandas==2.3.0
openpyxl==3.1.5
python-pptx==1.0.2
readability-lxml==0.8.4.1
lxml_html_clean==0.4.2
html-text==0.7.0
markdown==3.8.1
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0
xxhash==3.5.0
tika==3.1.0
openai==1.88.0
jieba==0.42.1
llama-index-core==0.10.68.post1
llama-index-postprocessor-dashscope-rerank-custom==0.1.0
dashscope==1.23.5
psycopg2-binary==2.9.10
bcrypt==4.3.0
fastapi-jwt==0.3.0
python-jose==3.5.0
authlib==1.6.0
colorlog==6.9.0
redis==5.0.1
aioredis==2.0.1
alembic==1.13.1
sqlalchemy==1.4.53